More code optimization

+++ Hazard/src/Hazard/Rendering/Buffers/ShaderDataType.h
@@ -30,17 +30,17 @@ namespace Hazard {
 	static uint32_t ShaderDataTypeSize(ShaderDataType type) {
 		switch (type)
 		{
-		case ShaderDataType::Float:    return 4;
-		case ShaderDataType::Float2:   return 4 * 2;
-		case ShaderDataType::Float3:   return 4 * 3;
-		case ShaderDataType::Float4:   return 4 * 4;
-		case ShaderDataType::Mat3:     return 4 * 3 * 3;
-		case ShaderDataType::Mat4:     return 4 * 4 * 3;
-		case ShaderDataType::Int:      return 4;
-		case ShaderDataType::Int2:     return 4 * 2;
-		case ShaderDataType::Int3:     return 4 * 3;
-		case ShaderDataType::Int4:     return 4 * 4;
-		case ShaderDataType::Bool:     return 1;
+		case ShaderDataType::Float:    return (uint32_t)4;
+		case ShaderDataType::Float2:   return (uint32_t)4 * 2;
+		case ShaderDataType::Float3:   return (uint32_t)4 * 3;
+		case ShaderDataType::Float4:   return (uint32_t)4 * 4;
+		case ShaderDataType::Mat3:     return (uint32_t)4 * 3 * 3;
+		case ShaderDataType::Mat4:     return (uint32_t)4 * 4 * 3;
+		case ShaderDataType::Int:      return (uint32_t)4;
+		case ShaderDataType::Int2:     return (uint32_t)4 * 2;
+		case ShaderDataType::Int3:     return (uint32_t)4 * 3;
+		case ShaderDataType::Int4:     return (uint32_t)4 * 4;
+		case ShaderDataType::Bool:     return (uint32_t)1;
 		}
 		return 0;
 	}
